requirements.txt
requests==2.26.0
pushbullet.py==0.12.0
schedule==1.1.0
python-dotenv==0.19.2